Introduced a breakpoint function to avoid assignment of variables to themselves for debugging purposes
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@12785 01821946-851b-0410-b7f4-f1477c70e57a

+++ effects/postprocessing.cpp
@@ -165,7 +165,7 @@ shaderManager.Set ("effectStrength", effectStrength);
 float viewport [4] = {
 	(float) gameData.render.frame.Left (), (float) gameData.render.scene.Top (), (float) gameData.render.scene.Width (), (float) gameData.render.scene.Height ()
 	};
-float screenSize [2] = { gameData.render.screen.Width (), gameData.render.screen.Height () };
+float screenSize [2] = { (float) gameData.render.screen.Width (), (float) gameData.render.screen.Height () };
 shaderManager.Set ("screenSize", screenSize);
 ogl.SetLighting (true);
 return true;
